AI-Memory-Vault_OLD/AI_Engine/requirements.txt
# ── Core Framework ────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# ── Embeddings & Vector Store ─────
langchain-huggingface>=0.1.0
langchain-chroma>=0.2.0
sentence-transformers>=3.0.0
chromadb>=0.5.0

# ── LLM Providers ────────────────
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0

# ── API Server ────────────────────
fastapi>=0.110.0
uvicorn>=0.27.0

# ── Messaging Gateway ────────────
httpx>=0.27.0

# ── Configuration ─────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0

# ── API Layer ─────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# ── Database ──────────────────────
SQLAlchemy>=2.0.0

# ── Hybrid Search ─────────────────
rank_bm25>=0.2.0

# ── MCP Server ────────────────────
mcp[cli]>=1.0.0
